# Speed up vanity salt mining by re-using a keccak hasher per worker thread