Reuse a keccak hasher across salt search attempts instead of creating a new one for each hash.

Speeds up vanity salt searches by at least 30% on my machine (AMD Ryzen 7 2700X Eight-Core with HT enabled).

NB: The following output depends on PR #190 which adds back functionality for `-a`, but for convenience, this PR does not include that commit.
```
jacob@Joshua:~/smartnode/rocketpool-cli (jms/vanity-2)$ ./build.sh 
jacob@Joshua:~/smartnode/rocketpool-cli (jms/vanity-2)$ ./rocketpool-cli-linux-amd64 minipool find-vanity-address -a 16 -p 0x123456

Running with 16 threads.
At salt 0x9641d0... 5.000144919s (1.96M salts/sec)
At salt 0x134c080... 10.000109476s (2.07M salts/sec)
At salt 0x1dfebb0... 15.002520997s (2.24M salts/sec)
Found on thread 12: salt 0x25bca1c = 0x1234565484a8f673752DC88DC78Acd7896845316
Finished in 19.016242865s

jacob@Joshua:~/smartnode/rocketpool-cli (jms/vanity-2)$ git stash save
Saved working directory and index state WIP on vanity-2: 4a095f75 Add missing cli parameter amount for find-vanity-address
jacob@Joshua:~/smartnode/rocketpool-cli (jms/vanity-2)$ ./build.sh 
jacob@Joshua:~/smartnode/rocketpool-cli (jms/vanity-2)$ ./rocketpool-cli-linux-amd64 minipool find-vanity-address -a 16 -p 0x123456

Running with 16 threads.
At salt 0x6d9aa0... 5.010013519s (1.43M salts/sec)
At salt 0xe032d0... 10.006862322s (1.5M salts/sec)
At salt 0x14fb540... 15.00679265s (1.46M salts/sec)
At salt 0x1c46dd0... 20.007108509s (1.52M salts/sec)
At salt 0x231a0e0... 25.018023519s (1.43M salts/sec)
Found on thread 12: salt 0x25bca1c = 0x1234565484a8f673752DC88DC78Acd7896845316
Finished in 27.680099122s
```