@jclapis 2022-02-18
Very nice, thank you for these optimizations!